chore: Select component refactoring - ColorSchemeControl - Iteration 5 (#15555)
* Enhance Select

* Transition Select to Antd

* Update test

* Fix Cypress

* Change name to aria-label

* Update Cypress search val

* Test Cypress selection
