+++ superset-frontend/cypress-base/cypress/integration/explore/visualizations/line.test.ts
@@ -73,13 +73,12 @@ describe('Visualization > Line', () => {
   it('should allow type to search color schemes', () => {
     cy.get('#controlSections-tab-display').click();
     cy.get('.Control[data-test="color_scheme"]').scrollIntoView();
-    cy.get('.Control[data-test="color_scheme"] input[type="text"]')
+    cy.get('.Control[data-test="color_scheme"] input[type="search"]')
       .focus()
-      .type('air{enter}');
-    cy.get('input[name="select-color_scheme"]').should(
-      'have.value',
-      'bnbColors',
-    );
+      .type('bnbColors{enter}');
+    cy.get(
+      '.Control[data-test="color_scheme"] .ant-select-selection-item > ul[data-test="bnbColors"]',
+    ).should('exist');
   });
 
   it('should work with adhoc metric', () => {
