+++ superset-frontend/spec/javascripts/explore/components/ColorScheme_spec.jsx
@@ -18,7 +18,7 @@
  */
 /* eslint-disable no-unused-expressions */
 import React from 'react';
-import { Select } from 'src/components/Select';
+import { Select } from 'src/components';
 import { getCategoricalSchemeRegistry } from '@superset-ui/core';
 import { styledMount as mount } from 'spec/helpers/theming';
 import ColorSchemeControl from 'src/explore/components/controls/ColorSchemeControl';
@@ -37,7 +37,7 @@ describe('ColorSchemeControl', () => {
     wrapper = mount(<ColorSchemeControl {...defaultProps} />);
   });
 
-  it('renders a Creatable', () => {
+  it('renders a Select', () => {
     expect(wrapper.find(Select)).toExist();
   });
 });
